Break out version modifications to separate extensions(schemascaffolder, dataobjectscaffolder). Remove debug code
